size now size_t

+++ socks6msg.cc
@@ -183,7 +183,7 @@ ssize_t S6M_Request_Packed_Size(const struct S6M_Request *req)
 	return err;
 }
 
-ssize_t S6M_Request_Pack(const struct S6M_Request *req, uint8_t *buf, int size)
+ssize_t S6M_Request_Pack(const struct S6M_Request *req, uint8_t *buf, size_t size)
 {
 	S6M_Error err;
 	
@@ -263,7 +263,7 @@ ssize_t S6M_AuthReply_Packed_Size(const struct S6M_AuthReply *authReply)
 	return err;
 }
 
-ssize_t S6M_AuthReply_Pack(const struct S6M_AuthReply *authReply, uint8_t *buf, int size)
+ssize_t S6M_AuthReply_Pack(const struct S6M_AuthReply *authReply, uint8_t *buf, size_t size)
 {
 	S6M_Error err;
 	
@@ -341,7 +341,7 @@ ssize_t S6M_OpReply_Packed_Size(const struct S6M_OpReply *opReply)
 	return err;
 }
 
-ssize_t S6M_OpReply_Pack(const struct S6M_OpReply *opReply, uint8_t *buf, int size)
+ssize_t S6M_OpReply_Pack(const struct S6M_OpReply *opReply, uint8_t *buf, size_t size)
 {
 	S6M_Error err;
 	
@@ -420,7 +420,7 @@ ssize_t S6M_PasswdReq_Packed_Size(const struct S6M_PasswdReq *pwReq)
 	return err;
 }
 
-ssize_t S6M_PasswdReq_Pack(const struct S6M_PasswdReq *pwReq, uint8_t *buf, int size)
+ssize_t S6M_PasswdReq_Pack(const struct S6M_PasswdReq *pwReq, uint8_t *buf, size_t size)
 {
 	S6M_Error err;
 	
@@ -488,7 +488,7 @@ ssize_t S6M_PasswdReply_Packed_Size(const struct S6M_PasswdReply *pwReply)
 	return UserPasswordReply::packedSize();
 }
 
-ssize_t S6M_PasswdReply_Pack(const struct S6M_PasswdReply *pwReply, uint8_t *buf, int size)
+ssize_t S6M_PasswdReply_Pack(const struct S6M_PasswdReply *pwReply, uint8_t *buf, size_t size)
 {
 	S6M_Error err;
